requirements.txt
hydra-core==1.3.2
imageio==2.35.1
matplotlib==3.9.2
numpy==2.1.2
omegaconf==2.3.0
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.14.1
